+++ spec/github/users/update_spec.rb
@@ -2,20 +2,21 @@
 
 require 'spec_helper'
 
-describe Github::Users, '#get' do
-  let(:user) { 'peter-murach' }
-  let(:request_path) { "/user?access_token=#{OAUTH_TOKEN}" }
+describe Github::Users, '#update' do
+  let(:key_id) { 1 }
+  let(:request_path) { "/user" }
 
   before {
     subject.oauth_token = OAUTH_TOKEN
-    stub_patch(request_path).with(user_params).
-      to_return(:body => body, :status => status,
+    stub_patch(request_path).with(:body => inputs,
+      :query => {:access_token => OAUTH_TOKEN}).
+    to_return(:body => body, :status => status,
       :headers => {:content_type => "application/json; charset=utf-8"})
   }
 
-  after { reset_authentication_for subject }
+  after { reset_authentication_for(subject) }
 
-  let(:user_params) {
+  let(:inputs) {
     {
       "name" => "monalisa octocat",
       "email" => "octocat@github.com",
@@ -32,23 +33,24 @@ describe Github::Users, '#get' do
     let(:status) { 200 }
 
     it "should create resource successfully" do
-      subject.update
-      a_patch(request_path).with(user_params).should have_been_made
+      subject.update inputs
+      a_patch(request_path).with(:body => inputs,
+        :query => {:access_token => OAUTH_TOKEN}).should have_been_made
     end
 
     it "should return the resource" do
-      user_resource = subject.update
+      user_resource = subject.update inputs
       user_resource.should be_a Hashie::Mash
     end
 
     it "should get the resource information" do
-      user_resource = subject.update
+      user_resource = subject.update inputs
       user_resource.login.should == 'octocat'
     end
   end
 
   it_should_behave_like 'request failure' do
-    let(:requestable) { subject.update }
+    let(:requestable) { subject.update inputs }
   end
 
 end # update