+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+0.8.3 (Dec 7, 2012)
+
+* Add oauth scopes listing method and helpers for reading scopes on a resource
+  accepted_oauth_scopes and oauth_scopes
+* Add say method call for printing octocat ASCII
+* Change issues listing to accept additional org and repo parameters.
+* Fix bug with milestones update method incorrect validation.
+* Change events listing to take issue_id as an optional parameter.
+
+0.8.2 (Dec 7, 2012)
+
+* Add Gitignore api.
+* Add listing of all repositories(a dump of every repository).
+* Add listing of all users (a dump of every user).
+* Add pull request comments listing in a repository.
+* Add issue comments listing in a repository.
+* Change unit tests for users api.
+* Update rspec, cucumber etc... dependencies
+
 0.8.1 (Nov 17, 2012)
 
 * Fix bug with validating options on Repository API create method.